pyproject.toml
[project]
name = "vea-backend"
version = "0.1.0"
description = "Backend service for managing xray/sing-box nodes and configurations"
authors = [{name = "Vea", email = "dev@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.24",
    "sqlmodel>=0.0.14",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "anyio>=3.6"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
